2012-03-11  Daniel Franke  <dfranke@users.sourceforge.net>
	* Imported r1238 from qwt/trunk.

+++ external/qwt/qwt_spline.cpp
@@ -2,7 +2,7 @@
  * Qwt Widget Library
  * Copyright (C) 1997   Josef Wilgen
  * Copyright (C) 2002   Uwe Rathmann
- * 
+ *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the Qwt License, Version 1.0
  *****************************************************************************/
@@ -14,7 +14,7 @@ class QwtSpline::PrivateData
 {
 public:
     PrivateData():
-        splineType(QwtSpline::Natural)
+        splineType( QwtSpline::Natural )
     {
     }
 
@@ -29,18 +29,18 @@ public:
     QPolygonF points;
 };
 
-static int lookup(double x, const QPolygonF &values)
+static int lookup( double x, const QPolygonF &values )
 {
 #if 0
 //qLowerBound/qHigherBound ???
 #endif
     int i1;
-    const int size = (int)values.size();
-    
-    if (x <= values[0].x())
-       i1 = 0;
-    else if (x >= values[size - 2].x())
-       i1 = size - 2;
+    const int size = values.size();
+
+    if ( x <= values[0].x() )
+        i1 = 0;
+    else if ( x >= values[size - 2].x() )
+        i1 = size - 2;
     else
     {
         i1 = 0;
@@ -49,12 +49,12 @@ static int lookup(double x, const QPolygonF &values)
 
         while ( i2 - i1 > 1 )
         {
-            i3 = i1 + ((i2 - i1) >> 1);
+            i3 = i1 + ( ( i2 - i1 ) >> 1 );
 
-            if (values[i3].x() > x)
-               i2 = i3;
+            if ( values[i3].x() > x )
+                i2 = i3;
             else
-               i1 = i3;
+                i1 = i3;
         }
     }
     return i1;
@@ -66,20 +66,20 @@ QwtSpline::QwtSpline()
     d_data = new PrivateData;
 }
 
-/*! 
+/*!
    Copy constructor
    \param other Spline used for initilization
 */
-QwtSpline::QwtSpline(const QwtSpline& other)
+QwtSpline::QwtSpline( const QwtSpline& other )
 {
-    d_data = new PrivateData(*other.d_data);
+    d_data = new PrivateData( *other.d_data );
 }
 
-/*! 
+/*!
    Assignment operator
    \param other Spline used for initilization
 */
-QwtSpline &QwtSpline::operator=( const QwtSpline &other)
+QwtSpline &QwtSpline::operator=( const QwtSpline & other )
 {
     *d_data = *other.d_data;
     return *this;
@@ -97,7 +97,7 @@ QwtSpline::~QwtSpline()
    \param splineType Spline type
    \sa splineType()
 */
-void QwtSpline::setSplineType(SplineType splineType)
+void QwtSpline::setSplineType( SplineType splineType )
 {
     d_data->splineType = splineType;
 }
@@ -116,36 +116,36 @@ QwtSpline::SplineType QwtSpline::splineType() const
 
   Depending on the value of \a periodic, this function
   will determine the coefficients for a natural or a periodic
-  spline and store them internally. 
-  
+  spline and store them internally.
+
   \param points Points
   \return true if successful
   \warning The sequence of x (but not y) values has to be strictly monotone
            increasing, which means <code>points[i].x() < points[i+1].x()</code>.
        If this is not the case, the function will return false
 */
-bool QwtSpline::setPoints(const QPolygonF& points)
+bool QwtSpline::setPoints( const QPolygonF& points )
 {
     const int size = points.size();
-    if (size <= 2) 
+    if ( size <= 2 )
     {
         reset();
         return false;
     }
 
     d_data->points = points;
-    
-    d_data->a.resize(size-1);
-    d_data->b.resize(size-1);
-    d_data->c.resize(size-1);
+
+    d_data->a.resize( size - 1 );
+    d_data->b.resize( size - 1 );
+    d_data->c.resize( size - 1 );
 
     bool ok;
     if ( d_data->splineType == Periodic )
-        ok = buildPeriodicSpline(points);
+        ok = buildPeriodicSpline( points );
     else
-        ok = buildNaturalSpline(points);
+        ok = buildNaturalSpline( points );
 
-    if (!ok) 
+    if ( !ok )
         reset();
 
     return ok;
@@ -181,10 +181,10 @@ const QVector<double> &QwtSpline::coefficientsC() const
 //! Free allocated memory and set size to 0
 void QwtSpline::reset()
 {
-    d_data->a.resize(0);
-    d_data->b.resize(0);
-    d_data->c.resize(0);
-    d_data->points.resize(0);
+    d_data->a.resize( 0 );
+    d_data->b.resize( 0 );
+    d_data->c.resize( 0 );
+    d_data->points.resize( 0 );
 }
 
 //! True if valid
@@ -194,18 +194,18 @@ bool QwtSpline::isValid() const
 }
 
 /*!
-  Calculate the interpolated function value corresponding 
+  Calculate the interpolated function value corresponding
   to a given argument x.
 */
-double QwtSpline::value(double x) const
+double QwtSpline::value( double x ) const
 {
-    if (d_data->a.size() == 0)
+    if ( d_data->a.size() == 0 )
         return 0.0;
 
-    const int i = lookup(x, d_data->points);
+    const int i = lookup( x, d_data->points );
 
     const double delta = x - d_data->points[i].x();
-    return( ( ( ( d_data->a[i] * delta) + d_data->b[i] ) 
+    return( ( ( ( d_data->a[i] * delta ) + d_data->b[i] )
         * delta + d_data->c[i] ) * delta + d_data->points[i].y() );
 }
 
@@ -213,10 +213,10 @@ double QwtSpline::value(double x) const
   \brief Determines the coefficients for a natural spline
   \return true if successful
 */
-bool QwtSpline::buildNaturalSpline(const QPolygonF &points)
+bool QwtSpline::buildNaturalSpline( const QPolygonF &points )
 {
     int i;
-    
+
     const QPointF *p = points.data();
     const int size = points.size();
 
@@ -226,58 +226,58 @@ bool QwtSpline::buildNaturalSpline(const QPolygonF &points)
 
     //  set up tridiagonal equation system; use coefficient
     //  vectors as temporary buffers
-    QVector<double> h(size-1);
-    for (i = 0; i < size - 1; i++) 
+    QVector<double> h( size - 1 );
+    for ( i = 0; i < size - 1; i++ )
     {
         h[i] = p[i+1].x() - p[i].x();
-        if (h[i] <= 0)
+        if ( h[i] <= 0 )
             return false;
     }
-    
-    QVector<double> d(size-1);
-    double dy1 = (p[1].y() - p[0].y()) / h[0];
-    for (i = 1; i < size - 1; i++)
+
+    QVector<double> d( size - 1 );
+    double dy1 = ( p[1].y() - p[0].y() ) / h[0];
+    for ( i = 1; i < size - 1; i++ )
     {
         b[i] = c[i] = h[i];
-        a[i] = 2.0 * (h[i-1] + h[i]);
+        a[i] = 2.0 * ( h[i-1] + h[i] );
 
-        const double dy2 = (p[i+1].y() - p[i].y()) / h[i];
-        d[i] = 6.0 * ( dy1 - dy2);
+        const double dy2 = ( p[i+1].y() - p[i].y() ) / h[i];
+        d[i] = 6.0 * ( dy1 - dy2 );
         dy1 = dy2;
     }
 
     //
     // solve it
     //
-    
+
     // L-U Factorization
-    for(i = 1; i < size - 2;i++)
+    for ( i = 1; i < size - 2; i++ )
     {
         c[i] /= a[i];
-        a[i+1] -= b[i] * c[i]; 
+        a[i+1] -= b[i] * c[i];
     }
 
     // forward elimination
-    QVector<double> s(size);
+    QVector<double> s( size );
     s[1] = d[1];
-    for ( i = 2; i < size - 1; i++)
-       s[i] = d[i] - c[i-1] * s[i-1];
-    
+    for ( i = 2; i < size - 1; i++ )
+        s[i] = d[i] - c[i-1] * s[i-1];
+
     // backward elimination
     s[size - 2] = - s[size - 2] / a[size - 2];
-    for (i = size -3; i > 0; i--)
-       s[i] = - (s[i] + b[i] * s[i+1]) / a[i];
+    for ( i = size - 3; i > 0; i-- )
+        s[i] = - ( s[i] + b[i] * s[i+1] ) / a[i];
     s[size - 1] = s[0] = 0.0;
 
     //
     // Finally, determine the spline coefficients
     //
-    for (i = 0; i < size - 1; i++)
+    for ( i = 0; i < size - 1; i++ )
     {
-        a[i] = ( s[i+1] - s[i] ) / ( 6.0 * h[i]);
+        a[i] = ( s[i+1] - s[i] ) / ( 6.0 * h[i] );
         b[i] = 0.5 * s[i];
-        c[i] = ( p[i+1].y() - p[i].y() ) / h[i] 
-            - (s[i+1] + 2.0 * s[i] ) * h[i] / 6.0; 
+        c[i] = ( p[i+1].y() - p[i].y() ) / h[i]
+            - ( s[i+1] + 2.0 * s[i] ) * h[i] / 6.0;
     }
 
     return true;
@@ -287,10 +287,10 @@ bool QwtSpline::buildNaturalSpline(const QPolygonF &points)
   \brief Determines the coefficients for a periodic spline
   \return true if successful
 */
-bool QwtSpline::buildPeriodicSpline(const QPolygonF &points)
+bool QwtSpline::buildPeriodicSpline( const QPolygonF &points )
 {
     int i;
-    
+
     const QPointF *p = points.data();
     const int size = points.size();
 
@@ -298,30 +298,30 @@ bool QwtSpline::buildPeriodicSpline(const QPolygonF &points)
     double *b = d_data->b.data();
     double *c = d_data->c.data();
 
-    QVector<double> d(size-1);
-    QVector<double> h(size-1);
-    QVector<double> s(size);
-    
+    QVector<double> d( size - 1 );
+    QVector<double> h( size - 1 );
+    QVector<double> s( size );
+
     //
     //  setup equation system; use coefficient
     //  vectors as temporary buffers
     //
-    for (i = 0; i < size - 1; i++)
+    for ( i = 0; i < size - 1; i++ )
     {
         h[i] = p[i+1].x() - p[i].x();
-        if (h[i] <= 0.0)
+        if ( h[i] <= 0.0 )
             return false;
     }
-    
+
     const int imax = size - 2;
     double htmp = h[imax];
-    double dy1 = (p[0].y() - p[imax].y()) / htmp;
-    for (i = 0; i <= imax; i++)
+    double dy1 = ( p[0].y() - p[imax].y() ) / htmp;
+    for ( i = 0; i <= imax; i++ )
     {
         b[i] = c[i] = h[i];
-        a[i] = 2.0 * (htmp + h[i]);
-        const double dy2 = (p[i+1].y() - p[i].y()) / h[i];
-        d[i] = 6.0 * ( dy1 - dy2);
+        a[i] = 2.0 * ( htmp + h[i] );
+        const double dy2 = ( p[i+1].y() - p[i].y() ) / h[i];
+        d[i] = 6.0 * ( dy1 - dy2 );
         dy1 = dy2;
         htmp = h[i];
     }
@@ -329,51 +329,51 @@ bool QwtSpline::buildPeriodicSpline(const QPolygonF &points)
     //
     // solve it
     //
-    
+
     // L-U Factorization
-    a[0] = qSqrt(a[0]);
+    a[0] = qSqrt( a[0] );
     c[0] = h[imax] / a[0];
     double sum = 0;
 
-    for( i = 0; i < imax - 1; i++)
+    for ( i = 0; i < imax - 1; i++ )
     {
         b[i] /= a[i];
-        if (i > 0)
-           c[i] = - c[i-1] * b[i-1] / a[i];
-        a[i+1] = qSqrt( a[i+1] - qwtSqr(b[i]));
-        sum += qwtSqr(c[i]);
+        if ( i > 0 )
+            c[i] = - c[i-1] * b[i-1] / a[i];
+        a[i+1] = qSqrt( a[i+1] - qwtSqr( b[i] ) );
+        sum += qwtSqr( c[i] );
     }
-    b[imax-1] = (b[imax-1] - c[imax-2] * b[imax-2]) / a[imax-1];
-    a[imax] = qSqrt(a[imax] - qwtSqr(b[imax-1]) - sum);
-    
+    b[imax-1] = ( b[imax-1] - c[imax-2] * b[imax-2] ) / a[imax-1];
+    a[imax] = qSqrt( a[imax] - qwtSqr( b[imax-1] ) - sum );
+
 
     // forward elimination
     s[0] = d[0] / a[0];
     sum = 0;
-    for( i = 1; i < imax; i++)
+    for ( i = 1; i < imax; i++ )
     {
-        s[i] = (d[i] - b[i-1] * s[i-1]) / a[i];
+        s[i] = ( d[i] - b[i-1] * s[i-1] ) / a[i];
         sum += c[i-1] * s[i-1];
     }
-    s[imax] = (d[imax] - b[imax-1] * s[imax-1] - sum) / a[imax];
-    
-    
+    s[imax] = ( d[imax] - b[imax-1] * s[imax-1] - sum ) / a[imax];
+
+
     // backward elimination
     s[imax] = - s[imax] / a[imax];
-    s[imax-1] = -(s[imax-1] + b[imax-1] * s[imax]) / a[imax-1];
-    for (i= imax - 2; i >= 0; i--)
-       s[i] = - (s[i] + b[i] * s[i+1] + c[i] * s[imax]) / a[i];
+    s[imax-1] = -( s[imax-1] + b[imax-1] * s[imax] ) / a[imax-1];
+    for ( i = imax - 2; i >= 0; i-- )
+        s[i] = - ( s[i] + b[i] * s[i+1] + c[i] * s[imax] ) / a[i];
 
     //
     // Finally, determine the spline coefficients
     //
     s[size-1] = s[0];
-    for ( i=0; i < size-1; i++)
+    for ( i = 0; i < size - 1; i++ )
     {
-        a[i] = ( s[i+1] - s[i] ) / ( 6.0 * h[i]);
+        a[i] = ( s[i+1] - s[i] ) / ( 6.0 * h[i] );
         b[i] = 0.5 * s[i];
-        c[i] = ( p[i+1].y() - p[i].y() ) 
-            / h[i] - (s[i+1] + 2.0 * s[i] ) * h[i] / 6.0; 
+        c[i] = ( p[i+1].y() - p[i].y() )
+            / h[i] - ( s[i+1] + 2.0 * s[i] ) * h[i] / 6.0;
     }
 
     return true;
